+++ sumd_logic_validator/sumd_logic_validator/__init__.py
@@ -1,3 +1,3 @@
 """SUMD Logic Validator logic runner."""
 
-__version__ = "0.3.52"
+__version__ = "0.3.53"